requirements.txt
streamlit
pandas
plotly
fpdf
Pillow